+++ src/File/Type/SpecializedFileType.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\AssetsBundle\File\Type;
+
+abstract class SpecializedFileType extends FileType
+{
+    /**
+     * Returns the extension that this file type supports.
+     */
+    abstract public static function supportsExtension () : string;
+}